more on ascending auctions (#45)
* more on ascending auctions

* ascending progress

* fix tests

* some forked changes

* simulation report progress

* simulation report messgae

* better utility logging

* anything else

+++ src/main/java/brown/user/main/library/ConfigRun.java
@@ -39,7 +39,7 @@ public class ConfigRun {
     this.config = config;
   }
 
-  public void run(Integer startingDelayTime, Integer simulationDelayTime,
+  public void run(Integer startingDelayTime, Double simulationDelayTime,
       Integer numSimulations) throws InstantiationException,
       IllegalAccessException, IllegalArgumentException,
       InvocationTargetException, InterruptedException {
